move ComputeJobExecutionType to internal_rpc
