+++ crates/compute_agent/src/commands/compute_job/queue_job.rs
@@ -1,6 +1,8 @@
 use crate::commands::compute_job::ComputeJobOpts;
 use anyhow::Result;
-use internal_rpc::{api::InternalRpcApiClient, client::InternalRpcClient};
+use internal_rpc::{
+    api::InternalRpcApiClient, client::InternalRpcClient, job_queue::ServiceJobType,
+};
 use service_config::ServiceConfig;
 
 /// Add a compute job to the server's job queue.
@@ -10,7 +12,7 @@ pub async fn run(opts: &ComputeJobOpts, config: &ServiceConfig) -> Result<uuid::
     let client = InternalRpcClient::new(config.rpc_socket_addr()?).await?;
     let job_uuid = client
         .0
-        .queue_job(&opts.cid, opts.job_type.to_owned())
+        .queue_job(&opts.cid, ServiceJobType::Compute(opts.job_type.to_owned()))
         .await?;
     println!("job UUID: {job_uuid:?}");
